+++ filtererror.go
@@ -1,32 +1,9 @@
 package dapper
 
-import (
-	"fmt"
-	"io"
-)
-
 // ErrorFilter is a [Filter] that formats implementations of [error].
-type ErrorFilter struct{}
-
-// Render writes a formatted representation of v to w.
-func (ErrorFilter) Render(
-	w io.Writer,
-	v Value,
-	c Config,
-	p FilterPrinter,
-) error {
-	e, ok := implements[error](v)
-	if !ok {
-		return ErrFilterNotApplicable
+func ErrorFilter(r Renderer, v Value) {
+	if e, ok := AsImplementationOf[error](v); ok {
+		r.WriteValue(v)
+		r.Print(" [%s]", e.Error())
 	}
-
-	if err := p.Fallback(w, c); err != nil {
-		return err
-	}
-
-	if _, err := fmt.Fprintf(w, " [%s]", e.Error()); err != nil {
-		return err
-	}
-
-	return nil
 }