+++ filterproto.go
@@ -1,0 +1,16 @@
+package dapper
+
+import "google.golang.org/protobuf/proto"
+
+// ProtoFilter is a [Filter] that formats implementations of [proto.Message].
+func ProtoFilter(r Renderer, v Value) {
+	if _, ok := AsImplementationOf[proto.Message](v); ok {
+		r.
+			WithModifiedConfig(
+				func(c *Config) {
+					c.OmitUnexportedFields = true
+				},
+			).
+			WriteValue(v)
+	}
+}