+++ filterstringer.go
@@ -1,10 +1,5 @@
 package dapper
 
-import (
-	"fmt"
-	"io"
-)
-
 // Stringer is an interface for types that produce their own Dapper
 // representation.
 type Stringer interface {
@@ -13,35 +8,21 @@ type Stringer interface {
 
 // StringerFilter is a [Filter] that formats implementations of
 // [dapper.Stringer].
-type StringerFilter struct{}
-
-// Render writes a formatted representation of v to w.
-func (StringerFilter) Render(
-	w io.Writer,
-	v Value,
-	c Config,
-	p FilterPrinter,
-) error {
-	stringer, ok := implements[Stringer](v)
+func StringerFilter(r Renderer, v Value) {
+	stringer, ok := AsImplementationOf[Stringer](v)
 	if !ok {
-		return ErrFilterNotApplicable
+		return
 	}
 
 	str := stringer.DapperString()
 	if str == "" {
-		return ErrFilterNotApplicable
+		return
 	}
 
 	if v.IsAmbiguousType() {
-		if err := p.WriteTypeName(w, v); err != nil {
-			return err
-		}
-
-		if _, err := w.Write(space); err != nil {
-			return err
-		}
+		r.WriteType(v)
+		r.Print(" ")
 	}
 
-	_, err := fmt.Fprintf(w, "[%s]", str)
-	return err
+	r.Print("[%s]", str)
 }