+++ internal/stream/indenter.go
@@ -5,14 +5,12 @@ import (
 	"io"
 )
 
-// DefaultIndent is the text used to make an DefaultIndent.
-var DefaultIndent = []byte("    ")
-
 // Indenter is an [io.Writer] that prefixes each line of text with a fixed
 // indent.
 type Indenter struct {
 	Target io.Writer
 	Indent []byte
+	Depth  int
 
 	indented bool
 }
@@ -54,11 +52,11 @@ func (w *Indenter) Write(data []byte) (int, error) {
 }
 
 func (w *Indenter) writeIndent() error {
-	indent := w.Indent
-	if len(indent) == 0 {
-		indent = DefaultIndent
+	for i := 0; i < w.Depth; i++ {
+		if _, err := w.Target.Write(w.Indent); err != nil {
+			return err
+		}
 	}
 
-	_, err := w.Target.Write(indent)
-	return err
+	return nil
 }